// UI/ClientApp/src/app/Core/Models/evaluationDetailsModel.ts
export class EvaluationDetailsModel {
  candidateId: number;
  practicalId: number;
  implemented: string;
  notImplemented: string;
  findings: string;
  reviewerComment: string;
  percentageComplete: number;
  remark: string;
  canProceed: number;
  grade: string;
  codeReview: number;
  canDoAttitude: number;
  communication: number;
}